z.12 Przenioslem klase Node do pliku StackList.java
